+++ src/main/java/org/opensearch/knn/common/KNNConstants.java
@@ -132,4 +132,9 @@ public class KNNConstants {
 
     // API Constants
     public static final String CLEAR_CACHE = "clear_cache";
+
+    // Radial search constants
+    public static final Float DEFAULT_LUCENE_RADIAL_SEARCH_TRAVERSAL_SIMILARITY_RATIO = 0.95f;
+    public static final String MIN_SCORE = "min_score";
+    public static final String MAX_DISTANCE = "max_distance";
 }