+++ src/main/java/org/opensearch/knn/index/query/RNNQueryFactory.java
@@ -5,6 +5,7 @@
 
 package org.opensearch.knn.index.query;
 
+import static org.opensearch.knn.common.KNNConstants.DEFAULT_LUCENE_RADIAL_SEARCH_TRAVERSAL_SIMILARITY_RATIO;
 import static org.opensearch.knn.common.KNNConstants.VECTOR_DATA_TYPE_FIELD;
 import static org.opensearch.knn.index.VectorDataType.SUPPORTED_VECTOR_DATA_TYPES;
 
@@ -118,7 +119,13 @@ public class RNNQueryFactory extends BaseQueryFactory {
         final float resultSimilarity,
         final Query filterQuery
     ) {
-        return new FloatVectorSimilarityQuery(fieldName, floatVector, resultSimilarity, filterQuery);
+        return new FloatVectorSimilarityQuery(
+            fieldName,
+            floatVector,
+            DEFAULT_LUCENE_RADIAL_SEARCH_TRAVERSAL_SIMILARITY_RATIO * resultSimilarity,
+            resultSimilarity,
+            filterQuery
+        );
     }
 
     /**
@@ -131,6 +138,12 @@ public class RNNQueryFactory extends BaseQueryFactory {
         final float resultSimilarity,
         final Query filterQuery
     ) {
-        return new ByteVectorSimilarityQuery(fieldName, byteVector, resultSimilarity, filterQuery);
+        return new ByteVectorSimilarityQuery(
+            fieldName,
+            byteVector,
+            DEFAULT_LUCENE_RADIAL_SEARCH_TRAVERSAL_SIMILARITY_RATIO * resultSimilarity,
+            resultSimilarity,
+            filterQuery
+        );
     }
 }